requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
python-dotenv>=1.0.0
selenium>=4.1.0
webdriver_manager>=4.0.0
pandas>=1.4.4
lxml>=4.9.1
pillow>=9.0.0
openpyxl>=3.1.0
tqdm>=4.64.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.2